// WCI/src/app/models/customtile.ts
export class Customtile {
    Id: number;
    Version: any = 0;
    Title: string;
    Type: any;
    Uri: string;
    Icon: any;
    ContentUrl: string;
    RequiresRefresh: boolean = false;
    RefreshInterval: number;
    ThumbnailOffsetX: number;
    ThumbnailOffsetY: number;
    ThumbnailScale: number   
}