// WCI/src/app/models/ModelEnums.ts
export enum RcMethod {
    NONE = 0,
    MSTSC = 1,
    VNC = 2
}

export enum RcThumbnailUpdate_Mode {
    NONE = 0,
    MODE_1 = 1,
    MODE_2 = 2
}

export enum DashboardMenuEnum {
   ACTIVE_ALERT = "Active Alert",
   THREE_D_VIEW = "3D View",
   PRIORITY_SAMPLES = "Priority Samples",
   REMOTE_CONTROL = "Remote Control",
   INTEGRATION = "Integrations",
   DIGITAL_CLOCK = "Digital Clock",
   ANALOG_CLOCK = "Analog Clock",
   CUSTOM_TIES = "Custom Tiles",
   MENU_LABEL = "menuLabel",
   REF_COMPONENT = "ref_component",
   RESET = "Reset",
   REMOVE_FROM_DASHBOARD = "Remove from Dashboard",
   ADD_TITLE = "Add Tile"
}